requirements.txt
fpdf2==2.7.7